+++ FfxivVr/vr/structs/ShaderMode.cs
@@ -1,0 +1,9 @@
+namespace FfxivVR;
+
+public enum ShaderMode
+{
+    Texture = 0,
+    Circle = 1,
+    InvertedAlpha = 2,
+    Fill = 3,
+}
