+++ FfxivVr/vr/structs/Vertex.cs
@@ -1,0 +1,17 @@
+using Silk.NET.OpenXR;
+
+namespace FfxivVR;
+
+public unsafe partial class Resources
+{
+    public struct Vertex
+    {
+        public Vector3f Position;
+        public Vector2f UV;
+        public Vertex(Vector3f position, Vector2f uv)
+        {
+            Position = position;
+            UV = uv;
+        }
+    }
+}
